+++ src/Controller/Admin/PhotosAlbumsController.php
@@ -39,9 +39,9 @@ class PhotosAlbumsController extends AppController
      */
     public function isAuthorized($user = null)
     {
-        //Only admins can delete albums
+        //Only admins and managers can delete albums
         if ($this->request->isDelete()) {
-            return $this->Auth->isGroup('admin');
+            return $this->Auth->isGroup(['admin', 'manager']);
         }
 
         return true;