+++ src/Controller/Component/AuthComponent.php
@@ -51,7 +51,10 @@ class AuthComponent extends CakeAuthComponent
             ],
             'authError' => __d('me_cms', 'You are not authorized for this action'),
             'authorize' => 'Controller',
-            'flash' => ['element' => 'MeTools.error'],
+            'flash' => [
+                'element' => 'MeTools.flash',
+                'params' => ['class' => 'alert-danger'],
+            ],
             'loginAction' => ['_name' => 'login'],
             'loginRedirect' => ['_name' => 'dashboard'],
             'logoutRedirect' => ['_name' => 'homepage'],