+++ src/Utility/StaticPage.php
@@ -22,10 +22,12 @@
  */
 namespace MeCms\Utility;
 
+use Cake\Cache\Cache;
 use Cake\Core\App;
 use Cake\Filesystem\Folder;
 use Cake\I18n\FrozenTime;
 use Cake\I18n\I18n;
+use Cake\ORM\Entity;
 use Cake\Utility\Inflector;
 use MeCms\Core\Plugin;
 
@@ -41,18 +43,24 @@ class StaticPage
      */
     protected static function paths()
     {
-        //Adds all plugins to paths
-        $paths = collection(Plugin::all())
-            ->map(function ($plugin) {
-                return collection(App::path('Template', $plugin))->first() . 'StaticPages';
-            })
-            ->filter(function ($path) {
-                return file_exists($path);
-            })
-            ->toList();
+        $paths = Cache::read('paths', 'static_pages');
 
-        //Adds APP to paths
-        array_unshift($paths, collection(App::path('Template'))->first() . 'StaticPages');
+        if (empty($paths)) {
+            //Adds all plugins to paths
+            $paths = collection(Plugin::all())
+                ->map(function ($plugin) {
+                    return collection(App::path('Template', $plugin))->first() . 'StaticPages';
+                })
+                ->filter(function ($path) {
+                    return file_exists($path);
+                })
+                ->toList();
+
+            //Adds APP to paths
+            array_unshift($paths, collection(App::path('Template'))->first() . 'StaticPages');
+
+            Cache::write('paths', $paths, 'static_pages');
+        }
 
         return $paths;
     }
@@ -87,13 +95,13 @@ class StaticPage
             $files = (new Folder($path))->findRecursive('^.+\.ctp$', true);
 
             foreach ($files as $file) {
-                $pages[] = (object)[
+                $pages[] = new Entity([
                     'filename' => pathinfo($file, PATHINFO_FILENAME),
                     'path' => rtr($file),
                     'slug' => self::slug($file, $path),
                     'title' => self::title(pathinfo($file, PATHINFO_FILENAME)),
                     'modified' => new FrozenTime(filemtime($file)),
-                ];
+                ]);
             }
         }
 
@@ -108,33 +116,48 @@ class StaticPage
      */
     public static function get($slug)
     {
-        //Sets the (partial) filename
-        $filename = implode(DS, af(explode('/', $slug)));
+        $locale = I18n::locale();
 
-        //Sets the filename patterns
-        $patterns = [sprintf('%s-%s', $filename, I18n::locale()), $filename];
+        //Sets the cache name
+        $cache = sprintf('page_%s_locale_%s', md5($slug), $locale);
 
-        //Checks if the page exists in APP
-        foreach ($patterns as $pattern) {
-            $filename = collection(App::path('Template'))->first() . 'StaticPages' . DS . $pattern . '.ctp';
+        $page = Cache::read($cache, 'static_pages');
 
-            if (is_readable($filename)) {
-                return 'StaticPages' . DS . $pattern;
-            }
-        }
+        if (empty($page)) {
+            //Sets the (partial) filename
+            $filename = implode(DS, af(explode('/', $slug)));
 
-        //Checks if the page exists in each plugin
-        foreach (Plugin::all() as $plugin) {
+            //Sets the filename patterns
+            $patterns = [$filename . '-' . $locale, $filename];
+
+            //Checks if the page exists in APP
             foreach ($patterns as $pattern) {
-                $filename = collection(App::path('Template', $plugin))->first() . 'StaticPages' . DS . $pattern . '.ctp';
+                $filename = collection(App::path('Template'))->first() . 'StaticPages' . DS . $pattern . '.ctp';
 
                 if (is_readable($filename)) {
-                    return sprintf('%s.%s', $plugin, 'StaticPages' . DS . $pattern);
+                    $page = 'StaticPages' . DS . $pattern;
+
+                    break;
                 }
             }
+
+            //Checks if the page exists in each plugin
+            foreach (Plugin::all() as $plugin) {
+                foreach ($patterns as $pattern) {
+                    $filename = collection(App::path('Template', $plugin))->first() . 'StaticPages' . DS . $pattern . '.ctp';
+
+                    if (is_readable($filename)) {
+                        $page = $plugin . '.' . 'StaticPages' . DS . $pattern;
+
+                        break;
+                    }
+                }
+            }
+
+            Cache::write($cache, $page, 'static_pages');
         }
 
-        return false;
+        return $page;
     }
 
     /**