+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # 2.x branch
 ## 2.16 branch
 ### 2.16.1
+* managers can delete photos and photos albums; 
 * fixed bug on redirect from the old address of the "contact us" form;
+* fixed bug in getting the active users list. Now it shows the full name;
 * fixed bug on `StaticPage::paths()`. Now it returns only existing paths;
+* `LoginRecorder` is now a component and returns the correct IP on localhost.
+    You must first set the user ID with the `config()` method and the `user`
+    value;
 * added `AppTable::beforeSave()`. It checks if the `created` property is an
     instance of `Time`;
+* static pages uses cache. `StaticPage::all()` returns an array of `Entity`;
 * improved `MeCms\Controller\Admin\LogController`, added `_read()` method and
     removed the `viewSerialized` action (the `view` action can use the 
     `view_as_serialized` template);